refactor: wrap CellUIFactory
