+++ src/main/java/com/holub/model/cell/Cell.java
@@ -1,9 +1,9 @@
-package com.holub.model;
+package com.holub.model.cell;
 
 import com.holub.life.Direction;
 import com.holub.life.Storable;
+import com.holub.model.Point;
 import com.holub.tools.Observable;
-import com.holub.ui.CellUI;
 
 /***
  * This interface is the basic unit that comprises a life board.
@@ -16,7 +16,6 @@ import com.holub.ui.CellUI;
 
 public interface Cell extends Observable {
 
-  CellUI getCellUI();
   boolean isUpdated();
   /**
    * Figure out the next state of the cell, given the specified neighbors.
